♻️ 코드 수정: 폴더 구조 전체적으로 변경
