+++ wizbuddy_front/src/assets/css/schedule/RegisterPage.css
