+++ wizbuddy_front/src/assets/css/schedule/WeeklyScheduleContainer.css
@@ -19,7 +19,7 @@
     align-items: center;
 }
 
-  .schedule-container {
+.schedule-container {
     max-width: 900px;
     width: 100%;
     margin-top: 20px;
@@ -104,7 +104,7 @@
     border-radius: 10px;
     text-align: center;
   }
-  
+
   .prev-next-button {
     padding-left: 20px;
     padding-right: 20px;
@@ -118,9 +118,3 @@
     width: 24px;
     height: 24px;
   }
-
-  .right-side {
-    width: 20%; 
-    background-color: #F3F7FA;
-    padding: 20px;
-}
